+++ app/apis/storage/client/providers/FireStorageClient.ts
@@ -1,4 +1,4 @@
-import { FileMetadata, RemoteFileUploadResDto, FileDto, RemoteFileDeleteResDto } from "../../../../types/dtos/remote.file.dtos";
+import { FileMetadata, RemoteFileUploadResDto, FileDto, RemoteFileDeleteResDto } from "../../../../types/dtos/remote-file.dtos";
 import { IFileStorageClient } from "../../../../types/interfaces/storage.client.interface";
 
 /** Storage Client for Provider Google Firestroage */
