+++ app/data/entities/cause.entity.ts
@@ -4,8 +4,8 @@
 
 import { Auditable } from "../../types/auditable";
 import { AuditableCauseDto, CauseDto } from "../../types/dtos/cause.dtos";
-import { FileDto } from "../../types/dtos/remote.file.dtos";
-import { EntityCreatedDto, EntityFetchedDto, EntityFetchedPageDto } from "../../types/dtos/server.message.dtos";
+import { FileDto } from "../../types/dtos/remote-file.dtos";
+import { EntityCreatedDto, EntityFetchedDto, EntityFetchedPageDto } from "../../types/dtos/server-message.dtos";
 import { Page } from "../../types/pagable";
 import CauseRepo from "../repos/cause.repo";
 import BaseEntity from "./base.entity";
