+++ app/data/repos/claim-evidence.repo.ts
@@ -8,8 +8,8 @@
 
 import { Auditable } from "../../types/auditable";
 import { ClaimEvidenceDto } from "../../types/dtos/claim-evidence";
-import { FileDto } from "../../types/dtos/remote.file.dtos";
-import { EntityFetchedDto, EntityFetchedPageDto, EntityCreatedDto, EntityUpdatedDto, EntityDeletedDto } from "../../types/dtos/server.message.dtos";
+import { FileDto } from "../../types/dtos/remote-file.dtos";
+import { EntityFetchedDto, EntityFetchedPageDto, EntityCreatedDto, EntityUpdatedDto, EntityDeletedDto } from "../../types/dtos/server-message.dtos";
 import ICRUDREPO from "../../types/interfaces/crud.repo.interface";
 import { Ownable } from "../../types/ownable";
 import { Page } from "../../types/pagable";
