+++ app/types/dtos/server-message.dtos.ts
