+++ app/types/interfaces/cause.entity.interface.ts
@@ -1,5 +1,5 @@
 import { Auditable } from "../auditable";
-import { AuditableFileMetadata, FileDto } from "../dtos/remote.file.dtos";
+import { AuditableFileMetadata, FileDto } from "../dtos/remote-file.dtos";
 
 export interface ICauseActions {
     $browser_UploadCauseAttachments(files: FileDto[]):Promise<[Auditable & FileDto]>;
