+++ tests/spec/app/data/entitiess/cause.entity.spec.ts
@@ -1,7 +1,7 @@
 import Cause from "../../../../../app/data/entities/cause.entity";
 import CauseRepo from "../../../../../app/data/repos/cause.repo";
 import { CauseDto } from "../../../../../app/types/dtos/cause.dtos";
-import { EntityUpdatedDto } from "../../../../../app/types/dtos/server.message.dtos";
+import { EntityUpdatedDto } from "../../../../../app/types/dtos/server-message.dtos";
 import { Auditable } from "../../../../../app/types/auditable";
 import { Ownable } from "../../../../../app/types/ownable";
